refactor: move queue to internal
